WarningView + extract common styles
of {Error|Warning|Success}View into ResultView.styles
