+++ src/renderer/components/shared/success/SuccessView.tsx
@@ -2,16 +2,17 @@ import React from 'react'
 
 import { ResultProps } from 'antd/lib/result'
 
+import * as RStyled from '../result/ResultView.styles'
 import * as Styled from './SuccessView.style'
 
 export type Props = Omit<ResultProps, 'icon'>
 
 export const SuccessView: React.FC<Props> = (props): JSX.Element => (
-  <Styled.Result
+  <RStyled.Result
     icon={
-      <Styled.IconWrapper>
+      <RStyled.IconWrapper>
         <Styled.Icon />
-      </Styled.IconWrapper>
+      </RStyled.IconWrapper>
     }
     {...props}
   />
