+++ src/renderer/components/shared/warning/WarningView.styles.ts
@@ -5,6 +5,6 @@ import { Icon as ResultIcon } from '../result/ResultView.styles'
 
 export const Icon = styled(ResultIcon)`
   svg {
-    fill: ${palette('error', 0)};
+    fill: ${palette('warning', 0)};
   }
 `
