+++ src/renderer/components/shared/warning/WarningView.tsx
@@ -3,11 +3,11 @@ import React from 'react'
 import { ResultProps } from 'antd/lib/result'
 
 import * as RStyled from '../result/ResultView.styles'
-import * as Styled from './ErrorView.styles'
+import * as Styled from './WarningView.styles'
 
 export type Props = Omit<ResultProps, 'icon'>
 
-export const ErrorView: React.FC<Props> = (props): JSX.Element => (
+export const WarningView: React.FC<Props> = (props): JSX.Element => (
   <RStyled.Result
     icon={
       <RStyled.IconWrapper>
