+++ src/renderer/components/shared/warning/index.ts
@@ -1,0 +1,1 @@
+export { WarningView } from './WarningView'
